Move question text to top of coordinate question inputs
The problem here is that the `QuestionFooterPresenter` for coordinate questions needs to be wrapped in a `CoordinateQuestionContext`. This means it needs a custom presenter. Previously this was included in `CoordinateQuestionPresenter` which keeps things neater, but because this is used as a header presenter, it was pushing the question text to the bottom of the inputs. This PR moves the footer into a separate presenter component to avoid this issue.